requirements.txt
# Django
Django==3.2.0

# django REST framework
djangorestframework == 3.12.0

# Debugging
ipdb==0.11

# Testing
pytest-django==4.4.0